requirements.txt
lxml==4.9.0
python-docx==0.8.11
SVGMath==0.3.3
ziafont==0.4
ziamath==0.6
